quick save - POLL CONTROLLER
added to docs, fixed some residual long ids, significant progress on poll controller, small inconsiquential changes

+++ WhiteboardAPI/WhiteboardAPI/Controllers/AssignmentControllers/AnnouncementController.cs
@@ -21,7 +21,7 @@ namespace WhiteboardAPI.Controllers
 		}
 
 		[HttpGet("{id}")]
-		public async Task<ActionResult<Announcement>> GetByID(long id)
+		public async Task<ActionResult<Announcement>> GetByID(int id)
 		{
 			var returnAnnouncement = await _context.Announcements.FindAsync(id);
